[#64] 포인트 충전 기능 구현하기 (#65)
* feat(point): 포인트 충전 기능 구현

* feat(point): 포인트 충전 기능 구현
-로그인한 사용자인지 확인
-세션을 통해 로그인 사용자 데이터 사용

+++ src/main/java/com/geonwoo/solokill/domain/payment/dto/PayRequest.java
@@ -1,8 +1,6 @@
 package com.geonwoo.solokill.domain.payment.dto;
 
-import java.math.BigDecimal;
-
 public record PayRequest(
-	BigDecimal amount
+	Long amount
 ) {
 }
